Fix some of the old and new conversion warnings in deflate*

+++ deflate_p.h
@@ -61,12 +61,12 @@ static inline int zng_tr_tally_dist(deflate_state *s, unsigned dist, unsigned ch
  * IN assertion: strstart is set to the end of the current match.
  */
 #define FLUSH_BLOCK_ONLY(s, last) { \
-    zng_tr_flush_block(s, (s->block_start >= 0L ? \
+    zng_tr_flush_block(s, (s->block_start >= 0 ? \
                    (char *)&s->window[(unsigned)s->block_start] : \
                    NULL), \
-                   (unsigned long)((long)s->strstart - s->block_start), \
+                   (unsigned long)((int)s->strstart - s->block_start), \
                    (last)); \
-    s->block_start = s->strstart; \
+    s->block_start = (int)s->strstart; \
     flush_pending(s->strm); \
 }
